# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Audio Guide 2026 API
  version: 1.13.0
  description: Canonical API for Audio Guide 2026. Includes fail-closed caching rules.
servers:
  - url: https://audiogid-api.vercel.app/v1
    description: Production Server
paths:
  /public/cities:
    get:
      summary: Get active cities
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      tags: [Public]
      responses:
        '200':
          description: List of active cities
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/City' }
        '304':
          $ref: '#/components/responses/NotModified'
  /public/catalog:
    get:
      summary: Get tours catalog for a city
      parameters:
        - name: city
          in: query
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfNoneMatch'
      tags: [Public]
      responses:
        '200':
          description: List of tours
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TourSnippet' }
        '304':
          $ref: '#/components/responses/NotModified'
  /public/poi/{poi_id}:
    get:
      summary: Get POI details with entitlement check
      parameters:
        - name: poi_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: city
          in: query
          required: true
          schema: { type: string }
        - name: device_anon_id
          in: query
          required: false
          schema: { type: string }
        - $ref: '#/components/parameters/IfNoneMatch'
      tags: [Public]
      responses:
        '200':
          description: POI details. Cache-Control depends on entitlement.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
            Vary: { $ref: '#/components/headers/Vary' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoiDetail' }
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          description: POI not found
...
components:
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: ETag for conditional GET
      schema: { type: string }
  headers:
    ETag:
      description: Version marker hash
      schema: { type: string }
    CacheControl:
      description: Caching policy (e.g. public, max-age=60 or private, no-store)
      schema: { type: string }
    Vary:
      description: Varied by Authorization for gated content
      schema: { type: string }
  responses:
    NotModified:
      description: Resource not changed
      headers:
        ETag: { $ref: '#/components/headers/ETag' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
        Vary: { $ref: '#/components/headers/Vary' }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    YookassaSignature:
      type: apiKey
      in: header
      name: X-Yookassa-Signature
  schemas:
    City:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name_ru: { type: string }
        is_active: { type: boolean }
        updated_at: { type: string, format: date-time }
    TourSnippet:
      type: object
      properties:
        id: { type: string, format: uuid }
        city_slug: { type: string }
        title_ru: { type: string }
        duration_minutes: { type: integer }
    PoiDetail:
      type: object
      properties:
        id: { type: string, format: uuid }
        title_ru: { type: string }
        description_ru: { type: string }
        lat: { type: number }
        lon: { type: number }
        preview_audio_url: { type: string, nullable: true }
        preview_bullets: { type: array, items: { type: string }, nullable: true }
        has_access: { type: boolean }
        narrations: { type: array, items: { $ref: '#/components/schemas/Narration' } }
        media: { type: array, items: { $ref: '#/components/schemas/Media' } }
        sources: { type: array, items: { $ref: '#/components/schemas/PoiSource' } }
        category: { type: string, nullable: true }
    Narration:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        locale: { type: string }
        duration_seconds: { type: number }
        transcript: { type: string, nullable: true }
    Media:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        media_type: { type: string }
        author: { type: string }
        source_page_url: { type: string }
        license_type: { type: string }
    PoiSource:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        url: { type: string, nullable: true }
...